+++ CHANGELOG.md
@@ -76,6 +76,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   [#435](https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox/-/issues/435)
 * Switching to Embedded Maps Service for extent maps
   [#256](https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox/-/issues/256)
+* Distribution option format detection now only uses media types
+  [#436](https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox/-/issues/436)
 
 ### Removed